// src/app/models/product.ts
export class product {
    "id": 0;
    "name": "string";
    "banner_img": "string";
    "cover_img": "string";
    "price": 0;
    "sale": 0;
    "description": "string";
    "availability": 0;
    "unit": "string";
    "storage_instructions": "string";
    "view": 0;
    "revenue": 0;
    "returned": 0;
    "status": 0;
    "seller_id": "string";
    "category_id": 0;
    "create_at": "2021-10-07";
    "update_at": "2021-10-07"
}